+++ test/useSwipe.spec.js
@@ -27,7 +27,7 @@ describe('useHorizontalSwipe', () => {
   it('should return the swipe state', () => {
     const { result } = renderHook(() => useHorizontalSwipe());
 
-    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alpha', 'count');
+    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alphaX', 'alphaY', 'count');
   });
 });
 
@@ -42,6 +42,6 @@ describe('useVerticalSwipe', () => {
   it('should return the swipe state', () => {
     const { result } = renderHook(() => useVerticalSwipe());
 
-    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alpha', 'count');
+    expect(result.current).to.be.an('object').that.has.all.keys('swiping', 'direction', 'alphaX', 'alphaY', 'count');
   });
 });